+++ src/render/dom/render.ts
@@ -3,7 +3,7 @@ import { MotionProps } from "../../motion/types"
 import { HTMLVisualElement } from "./HTMLVisualElement"
 import { filterProps } from "./utils/filter-props"
 import { buildHTMLProps } from "./utils/build-html-props"
-import { buildSVGProps } from "./utils/build-svg-attrs"
+import { buildSVGProps } from "./utils/build-svg-props"
 import { SVGVisualElement } from "./SVGVisualElement"
 import { isSVGComponent } from "./utils/is-svg-component"
 
@@ -16,6 +16,15 @@ export function render<Props>(
     // is a custom component pass along everything provided to it.
     const forwardedProps =
         typeof Component === "string" ? filterProps(props) : props
+
+    /**
+     * Every render, empty and rebuild the animated values to be applied to our Element.
+     * During animation these data structures are used in a mutable fashion to reduce
+     * garbage collection, but between renders we can flush them to removed values
+     * that might have been taken out of the provided props.
+     */
+    visualElement.clean()
+    visualElement.build()
 
     // Generate props to visually render this component
     const visualProps = isSVGComponent(Component)